+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.1.0
 
+- Add `borderless` and `tagRenderer` props to AutocompleteTagBuilder
 - Remove refocus of component(Autocomplete) on every value change, remove corresponding test
 - Fix TagBuilder not deleting tags properly with backspace (#223)